+++ addon/nodes/doc.ts
@@ -1,5 +1,31 @@
 import { NodeSpec } from 'prosemirror-model';
 
-export const doc: NodeSpec = {
-  content: 'block+',
+interface DocumentConfig {
+  defaultLanguage?: string;
+  content?: string;
+}
+
+// Note: the `doc` node-spec does not have any parsing rules, as the parsing of the doc node is done in the `initalize` method
+// of the `SayController` class.
+export const docWithConfig: (config?: DocumentConfig) => NodeSpec = ({
+  defaultLanguage = 'nl-BE',
+  content = 'block+',
+} = {}) => {
+  return {
+    content,
+    attrs: {
+      lang: {
+        default: defaultLanguage,
+      },
+    },
+    toDOM(node) {
+      return [
+        'div',
+        { lang: node.attrs.lang as string, 'data-say-document': true },
+        0,
+      ];
+    },
+  };
 };
+
+export const doc = docWithConfig();